Implementation First Draft
Here is the first draft of our Jupyter Notebook implementation of Grover's Algorithm. The article used as a reference for this circuit used a slightly different approach than we did in our previous homeworks, so some of the implementation may need to be redone to better fit how we know to get this working.