Adhering to Linting and Corrections to Functions
Here, I've fixed the linting throughout the whole notebook, except for the API token. Also, I've redefined two of the functions to properly apply the gates that we want them to apply. I've also changed it so that the gates are all applied with the functions that were defined. This shows the process a little nicer and cleaner, and adds capability for more easy generalization.